Move adversarial walker environments to one file
